refactor(agentic-loop): Optimize tool call counting and budget control logic Fix tool call counting logic to only count actually executed tool calls Update budget control threshold checks with correct equality semantics Remove pre-increment logic to ensure rejected or not-found calls do not consume budget Fix counting issues in parallel tool execution Add conditional checks on execution status to correctly update metrics Refactor comparison operators in budget check function to match expected behavior Provide more accurate documentation for tool call counting

+++ crates/lc-core/src/http.rs
@@ -740,7 +740,7 @@ fn random_entropy() -> u64 {
     // `rand` dependency through the core crate.
     SystemTime::now()
         .duration_since(UNIX_EPOCH)
-        .map(|d| (d.as_secs() ^ d.subsec_nanos() as u64) as u64)
+        .map(|d| d.as_secs() ^ d.subsec_nanos() as u64)
         .unwrap_or(0)
 }
